ath6kl: Add power gpio reset feature into sdio driver for suspend
This change adds pm suspend callbacks in order to trigger a gpio line
to push the PWD_L reset/power line low on suspend. This puts the chip
into the lowest power state on suspend. On resume, it releases the line,
allowing the chip to boot. Slower, but provides a clean reset of the
chip and recovery from standby.

Bug: 4651