atmel_usba_udc: Add ablity to suspend system and wakeup based on bus activity flags
This change gives the USB gadget driver the capability to initiate a system-wide
power management suspend when the USB bus goes to an idle, and to wake the system from
suspend when activity returns to the bus.

The following features have been added:
* Ability to suspend the system when the DET_SUSPD interrupt is triggered
* Ability to wake the system when suspended when the WAKE_UP interrupt is triggered
* sysfs attributes have been added to enable and disable the feature as well as select
  the type of suspend (standby or mem) to do.
* Added new DBG_PM debugging print class

I also changed some logic in the usba_udc_irq ISR. Seems that the wakeup flag is always
set when the suspend flag is set. Had to make it so that we don't do the wakeup state when
we try to suspend.

Bug: 4571