Save the firmware string for later retrieval
The original catalyst for the bug was CCX status needed to be determined
by the cli on the WB40.  The necessitated getting the firmware string
from the driver.  This change gives the same capability to the 45
driver.

The firmware version is displayed in sdc_cli version.

The driver version is set to 3.5.0.0 to sync with latest SDK

BZ 4670