Allow connect to AP with multiple ciphers
The WB45 was not connecting to an AP configured for WPA2 but allowing
both AES and TKIP encryption.  When the WB45 was configured for wpa2-aes,
it would not connect to this mixed mode AP configuration.  It should
connect using WPA2 with AES as the unicast cipher and TKIP for the group
cipher but the WB45 would not attempt to connect (probes sent and probe responses, but never attempted to associate).

By or'ing in the ignore wpax cipher flag resolves the issue.

BZ5014