laird: ath6kl: use add_krk_cmd size from firmware .h files
The size of the add_krk_cmd struct in both the 6003 and 6004 firmware
files is 26 but the driver had it set to 16.  The 6003 firmware did not
have issue with this but the 6004 firmware will fail to allow CCKM fast
roaming with the smaller size.  This change set the size to 26 to match
the firmware size.

bug 8853