laird: ath6kl: ignore WMI_TXE_NOTIFY_EVENTID based on fw capability flags
The 6004 redefines the WMI_TXE_NOTIFY_EVENTID event number and sends the
new event frequently. However it doesn't have the tx-err-notify feature
and thus this firmware capability flag isn't set on the firmware package.
By guarding the processing of this event by the same method we guard the
sending of the WMI_SET_TXE_NOTIFY_CMDID command, we can ignore the spurious
event that we don't know how to process yet.

Without this change we call cfg80211_cqm_txe_notify() with possibly bad
data.

Bug: 9049