laird: ath6kl: sme_state shortcut to SME_DISCONNECTED removed
When an NL80211_DISCONNECT is sent to cfg80211, the driver's cfg80211
disconnect function sets the sme_state to SME_DISCONNECTED before receiving
a WMI_DISCONNECT_EVENT from the firmware.  This was causing the supplicant,
SDK events, and the SDK status function to not see that the connection is
currently disconnected becuase the driver fails to notify cfg80211 of the
successful disconnect.  This also caused the supplicant to fail to
reconnect after standby.

Bug: 6050 9085