laird: brcmfmac: Add bcm43353 sdio WiFi device
The bcm43353 is a 802.11 a/b/g/n/ac WiFi and BT 4.1 chip where the WiFi
component is essentially a 4339. The device answers with an SDIO device
ID of 0x4339 and uses the 4339 firmware.

This commit adds the 43353 so it can be identified and bound to the
brcmfmac driver when the device identifies itself on the SDIO bus.

It was tested with a Laird Sterling LWB5 and works as expected.

Bug: 9699