laird: ath6kl_usb: add WLAN_PWD_L gpio control
This adds the same style GPIO control for the WLAN_PWD_L reset pin as we
already have on the sdio driver. This is the first step in getting control
of the pin for suspend/resume and firmware recovery.

I had to replace the simplfied macro-based module_usb_driver() setup with
a manual driver init and setup as I've added specialized setup for the GPIO.

Bug: 10255