laird: mwlwifi: add sdio support for sdio 8897 and 8997
This builds on the previous change splitting out the interface
drivers and adds a driver for SDIO chips, namely the 8897 and
8997.

The bulk of the mwlwifi_sdio code came from Marvel's vendor
driver and was largely retained intact, other than a few changes
for some obvious problems, style issues and fixes to modernize the
driver. Also a number of functions were reordered to avoid forward
declarations.

This does function properly as tested on a WB50-RDVK with a Laird
SD60 (88W8997).

Bug: 10593