laird: mwlwifi: Remove setting self MAC on connect to AP
This chunk of code is questionable for a station - why tell the firmware
our own MAC for the stn_ap db? There's no obvious reason to do so. Testing
with this chunk of code removed showed throughput numbers between 2x and 3x
as with it there. So removing this chunk.

It's possible this code may need to be here to fulfill some firmware
requirement, or it might need to be utilized in other modes. We'll kill
it for now, but later we might choose to revert this commit.

Bug: 10604