laird: brcmfmac - enable MMC detection after driver unloads or fails
A previous commit from upstream set the MMC_CAP_NONREMOVABLE bit as part
of a larger WOW enhancement. This causes an issue if the driver is to be
unloaded and reloaded as the state of the MMC is left so that card
detect is disabled.

Enable MMC's card detect when the driver is unloaded or fails to load.