laird: allow key to be injected to a keyring for encrypted fs use
The IG needs a key to be passed from encrypted storage in uboots
.itb which is then loaded into a keyring for use with the encrypted
RW filesystem.
The key can be added to the running users session with:

keyctl search %:_builtin_fs_keys logon fscrypt:ffffffffffffffff
%:_uid_ses.0

With the key not present any filename will be encrypted and contents
will fail to read with key not present.

Bug 12815