Fix RS-485 implementation on Atmel serial driver
The Linux RS-485 IOCTLs are intended to disable Rx operation on the
serial port when Tx is in progress (since half-duplex hardware
will echo back).  The Atmel implementation was incorrect, and re-
enabled Rx on DMA completion, which allowed a few stray chars
since there was still data in the THR.  The fix is to use the
TXEMPTY IRQ after DMA is complete to enable Rx.

Bug 13499