Laird: WB45/50 - Disable runtime power management autosuspend on sdio radio interface
Disable runtime power management autosuspend on the SDIO interface used by the radio.
This forces the SDIO clock to continue running and is required in order for the host to immediately detect interrupts generated by the radio.
This change fixes issues such as high ping latencies observed in some scenarios.

Bug: 13859