bluetooth: Slave requests link latency, timeout change
Change the logic to request an update to the link parameters to include
the latency and supervisory timeout that were present in debugfs when
the connection was established.  Prior logic was to request the change
only based on the connection interval, and always use the latency and
timeout established by the master.

Bug 13938