brcmfmac: TWT: Add support to handle the async TWT events from the Firmware
Introduce a lock protected Doubly Linked List to maintain the TWT sessions
on each interface. This list helps later in avoiding an expensive TWT iovar
call to the Firmware, that is likely to get returned with a failure.

Register a handler for the TWT Setup and TWT Teardown events on verifying
that the Firmware supports TWT iovar. Later when receiving an asynchronous
TWT event as a response to the TWT Setup and TWT Teardown requests sent to
the firmware, update the TWT session list of the interface appropriately.

Do a lookup on the TWT session list before attempting to initiate either a
TWT Setup request or a TWT Teardown request to the Firmware. This helps in
deciding whether to pass the TWT Operation request to the Firmware or not.
For Example, when an iTWT session with Flow ID 3 is already active, any
further iTWT setup requests with same Flow ID on the same interface should
get rejected instead of passing it to the firmware using an expensive
operation. Likewise, some of the invalid TWT session setup and teardown
requests can be returned with a failure from the host driver itself.

To avoid maintaining stale session entries, cleanup the TWT session list on
receiving a disconnect from cfg80211 layer or when TWT Teardown event with
"Teardown All TWT" is received.

Signed-off-by: Gokul Sivakumar <gokulkumar.sivakumar@infineon.com>