brcmfmac: TWT: cleanup stale session entries after receiving Deauth from AP
Firmware sends TWT Setup & Teardown MAC events to Driver on the completion
of the Session Setup & Teardown tasks. On receiving these events, driver
updates its state machine.

If the disconnection is initiated locally, the session cleanup is handled
implicitly. However when the Deauth is initiated by the AP, the firmware
silently tears down the TWT session and disconnects without sending the TWT
Teardown event to Driver. And the driver continues to maintain the stale
TWT sessions. Later if the user attempts to initiate a TWT session with the
same Flow ID in the stale entry, the request gets rejected.

To avoid this, handle this corner case and cleanup the stale TWT sessions
while bringing down the STA interface after handling Deauth.

Signed-off-by: Gokul Sivakumar <gokulkumar.sivakumar@infineon.com>