brcmfmac: Fix SDIO bus access in SDIO sleep state
Issue: SDIO write/read operation is happening when bus is in sleep state.
Cause: There was a corner case when watchdog thread is trying to claim host
ISR is occurred and dpc is triggered. Watchdog thread is not aware of this
change and once the watchdog thread context execution is resumed it
tries to put bus in sleep state, which is causing the dpc execution of
SDIO read/write in bus sleep state.
Fix: Added additional check after host is claimed by watchdog thread. If
DPC is triggered or running skip SDIO bus sleep.

Signed-off-by: avishad verma <avishad.verma@infineon.com>