brcmfmac: enable remote wakeup before USB suspend, if WOWL configs are set
When the host system attached with USB WLAN device goes to suspend state,
remote wakeup was supposed to be enabled for the device, to allow waking
the system based on the configured WOWL triggers.

Currently, the remote wakeup was enabled in the device driver's USB suspend
routine, which would not take effect until the next suspend, so the WOWL
triggers are ineffective in the current suspend. And in the device driver's
USB resume routine, the remote wakeup is disabled, so the remote wakeup
that is supposed to be effective in the next routine also got avoided.
This cycle makes the remote wakeup ineffective for all the received WOWL
trigger instances.

So before USB suspend, enable remote wakeup for USB device while setting
the WOWL triggers in the USB Device firmware.

Signed-off-by: Gokul Sivakumar <gokulkumar.sivakumar@infineon.com>