brcmfmac: fix get_assoc_ies misused endian untranslated variable.
Commit 264ae44ae2c ("brcmfmac: External roam time enhancements") correctly
introduces endian conversion for "assoc_info->flags" in the function
brcmf_get_assoc_ies(), but fails to use the converted value in the flags
variable to check for REASSOC bit. But this does not have an impact.

Because brcmfmac does not traverse this code path to query the firmware
for collecting the ASSOC Request & Response IE, as the ASSOC IE is already
collected from the firmware via the registered asynchronous firmware events
BRCMF_E_ASSOC_REQ_IE(87) and BRCMF_E_ASSOC_RESP_IE(88).

Signed-off-by: Carter Chen <carter.chen@infineon.com>